Map widgets using data from destination
